+++ Editor/Component/ComponentEditorWindow.cs
@@ -4,31 +4,51 @@ using UnityEngine;
 using UnityObject = UnityEngine.Object;
 
 namespace HananokiEditor.CustomHierarchy {
-	public class Window_ComponentEditor : HNEditorWindow<Window_ComponentEditor> {
-		UnityObject m_component;
-		bool inited;
+	public class ComponentEditorWindow : HNEditorWindow<ComponentEditorWindow> {
+
 		Vector2 m_scroll;
 		public Editor m_currentEditor;
+
+		static UnityObject s_component;
+		static ComponentEditorWindow s_window;
+
 
 
 		/////////////////////////////////////////
 		public static void Open( UnityObject component ) {
-			var window = EditorWindowUtils.ShowWindow<Window_ComponentEditor>();
-			window.m_component = component;
+			s_component = component;
+			if( s_window == null ) {
+				s_window = EditorWindowUtils.ShowWindow<ComponentEditorWindow>();
+			}
+			else {
+				s_window.Init();
+				s_window.Repaint();
+				s_window.Focus();
+			}
+		}
+
+
+		/////////////////////////////////////////
+		void OnEnable() {
+			Init();
+		}
+
+
+		/////////////////////////////////////////
+		void OnDestroy() {
+			s_window = null;
 		}
 
 
 		/////////////////////////////////////////
 		void Init() {
-			SetTitle( m_component.GetType().Name, m_component.GetCachedIcon() );
-			m_currentEditor = ComponentEditorCache.Get( m_component );
-			inited = true;
+			SetTitle( s_component.GetType().Name, s_component.GetCachedIcon() );
+			m_currentEditor = ComponentEditorCache.Get( s_component );
 		}
 
 
 		/////////////////////////////////////////
 		public override void OnDefaultGUI() {
-			if( !inited ) Init();
 
 			EditorGUIUtility.hierarchyMode = true;
 			EditorGUIUtility.wideMode = true;
@@ -40,8 +60,6 @@ namespace HananokiEditor.CustomHierarchy {
 			using( var sc = new GUILayout.ScrollViewScope( m_scroll ) )
 			using( new GUILayoutScope( 16, 4 ) ) {
 				m_scroll = sc.scrollPosition;
-				//m_currentEditor?.DrawDefaultInspector();
-
 				m_currentEditor?.OnInspectorGUI();
 			}
 		}