+++ Editor/Function/CommandBar.cs
@@ -1,7 +1,6 @@
 ﻿using UnityEngine;
 using HananokiEditor.Extensions;
 using System;
-using System.Linq;
 using UnityEditor;
 
 #if UNITY_2019_1_OR_NEWER
@@ -10,7 +9,7 @@ using UnityEngine.UIElements;
 
 
 
-using E = HananokiEditor.CustomHierarchy.SettingsEditor;
+using E = HananokiEditor.CustomHierarchy.EditorPref;
 
 
 namespace HananokiEditor.CustomHierarchy {
@@ -53,24 +52,41 @@ namespace HananokiEditor.CustomHierarchy {
 			s_sceneHierarchy?.AddIMGUIContainer( s_IMGUIContainer );
 
 #if UNITY_2019_1_OR_NEWER
-			IMGUIContainer con = (IMGUIContainer) s_IMGUIContainer;
+			IMGUIContainer imgui = (IMGUIContainer) s_IMGUIContainer;
+
+			imgui.style.position = Position.Absolute;
 
 			if( UnitySymbol.UNITY_2019_3_OR_NEWER ) {
-				con.style.height = 20;
-				con.style.width = 26 * 2 + 36 * 2;
-				con.style.marginLeft = 36;
+				imgui.style.height = 20;
+				if( ExternalPackages.SelectionHistory.enabled ) {
+					imgui.style.width = 26 * 2 + 36 * 1;
+				}
+				else {
+					imgui.style.width = 36 * 2;
+				}
+				imgui.style.left = 36;
 			}
 			else if( UnitySymbol.UNITY_2019_2_OR_NEWER ) {
-				con.style.height = 18;
-				con.style.width = 28;
-				con.style.marginLeft = 50;
-				con.style.marginTop = -2;
+				imgui.style.height = 18;
+				if( ExternalPackages.SelectionHistory.enabled ) {
+					imgui.style.width = 26 * 2 + 40 * 1;
+				}
+				else {
+					imgui.style.width = 40 * 1;
+				}
+				imgui.style.top = -2;
+				imgui.style.left = 49;
 			}
 			else {
-				con.style.top = -1;
-				con.style.height = 16;
-				con.style.width = 28;
-				con.style.marginLeft = 50;
+				imgui.style.height = 16;
+				if( ExternalPackages.SelectionHistory.enabled ) {
+					imgui.style.width = 26 * 2 + 40 * 1;
+				}
+				else {
+					imgui.style.width = 40 * 1;
+				}
+				imgui.style.top = -1;
+				imgui.style.left = 49;
 			}
 #endif
 			//if( !UnitySymbol.UNITY_2019_3_OR_NEWER ) {
@@ -84,28 +100,35 @@ namespace HananokiEditor.CustomHierarchy {
 			GUILayout.BeginArea( new Rect( 0, 0, s_sceneHierarchy.position.width, 20 ) );
 			ScopeHorizontal.Begin();
 			GUILayout.Space( 1 );
-			//			"Timeline.MixLeft"
-			if( HGUIToolbar.Button( EditorIcon.tab_prev ) ) Undo.PerformUndo();
-			if( HGUIToolbar.Button( EditorIcon.tab_next ) ) Undo.PerformRedo();
 
-			HGUIToolbar.DropDown( EditorIcon.settings, () => {
-				var m = new GenericMenu();
-				m.AddItem( S._UnhideGameObject, _UnhideGameObject );
-				m.DropDown( HEditorGUI.lastRect );
-			} );
-			HGUIToolbar.DropDown( EditorIcon.search_icon, () => {
-				var m = new GenericMenu();
-				m.AddItem( "None", _Filter, null );
-				var ll = E.i.m_componentHandlerData.Where( x => x.type != null ).OrderBy( x => x.type.Name ).Where( x => x.search );
-				if( 0 < ll.Count() ) {
-					m.AddSeparator();
-					foreach( var p in ll ) {
-						m.AddItem( p.type.Name, _Filter, p );
-					}
-				}
+			if( ExternalPackages.SelectionHistory.enabled ) {
+				ScopeDisable.Begin( !ExternalPackages.SelectionHistory.hasPrev );
+				if( HGUIToolbar.Button( EditorIcon.tab_prev ) ) ExternalPackages.SelectionHistory.Prev();
+				ScopeDisable.End();
+				ScopeDisable.Begin( !ExternalPackages.SelectionHistory.hasNext );
+				if( HGUIToolbar.Button( EditorIcon.tab_next ) ) ExternalPackages.SelectionHistory.Next();
+				ScopeDisable.End();
+			}
+			HGUIToolbar.DropDown( EditorIcon.AnimationFilterBySelection, () => {
+				//var m = new GenericMenu();
+				//m.AddItem( S._UnhideGameObject, _UnhideGameObject );
+				//m.DropDown( HEditorGUI.lastRect );
+				PopupContentHideSelect.Show( HEditorGUI.lastRect );
+				//SceneVisibilityManager.instance.HideAll();
+			}, GUILayout.Width( 38 ) );
+			//HGUIToolbar.DropDown( EditorIcon.search_icon, () => {
+			//	var m = new GenericMenu();
+			//	m.AddItem( "None", _Filter, null );
+			//	var ll = E.i.m_componentHandlerData.Where( x => x.type != null ).OrderBy( x => x.type.Name ).Where( x => x.search );
+			//	if( 0 < ll.Count() ) {
+			//		m.AddSeparator();
+			//		foreach( var p in ll ) {
+			//			m.AddItem( p.type.Name, _Filter, p );
+			//		}
+			//	}
 
-				m.DropDown( HEditorGUI.lastRect );
-			} );
+			//	m.DropDown( HEditorGUI.lastRect );
+			//} );
 
 			ScopeHorizontal.End();
 			GUILayout.EndArea();
@@ -118,20 +141,20 @@ namespace HananokiEditor.CustomHierarchy {
 			}
 		}
 
-		static void _Filter( object context ) {
-			var p = (ComponentHandlerData) context;
-			if( p != null ) {
-				//	filter = p.type.Name;
-				//	mode = SearchableEditorWindow.SearchMode.Type;
-				//	//filter = s.filter.Split('/').Last();
-				//	//mode = s.searchMode;
-				//SceneHierarchyUtils.SetSearchFilter( filter, mode );
-				SceneModeUtility.SearchForType( p.type );
-			}
-			else {
-				SceneModeUtility.SearchForType( null );
-			}
-		}
+		//static void _Filter( object context ) {
+		//	var p = (ComponentHandlerData) context;
+		//	if( p != null ) {
+		//		//	filter = p.type.Name;
+		//		//	mode = SearchableEditorWindow.SearchMode.Type;
+		//		//	//filter = s.filter.Split('/').Last();
+		//		//	//mode = s.searchMode;
+		//		//SceneHierarchyUtils.SetSearchFilter( filter, mode );
+		//		SceneModeUtility.SearchForType( p.type );
+		//	}
+		//	else {
+		//		SceneModeUtility.SearchForType( null );
+		//	}
+		//}
 
 	}
 }