+++ Editor/Function/HideSelect/PopupContentHideSelect.cs
@@ -1,4 +1,9 @@
-﻿#pragma warning disable 649
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+
+#pragma warning disable 649
 
 using HananokiRuntime;
 using UnityEditor;
@@ -7,17 +12,14 @@ using UnityObject = UnityEngine.Object;
 
 namespace HananokiEditor.CustomHierarchy {
 
-	public class PopupContent_Component : PopupWindowContent {
+	public class PopupContentHideSelect : PopupWindowContent {
 
-		static PopupContent_Component s_content;
-		static UnityObject s_component;
+		static PopupContentHideSelect s_content;
 
-		public Editor m_currentEditor;
 		Vector2 m_scroll;
 
 		/////////////////////////////////////////
-		public static void Show( Rect rect, UnityObject component ) {
-			s_component = component;
+		public static void Show( Rect rect ) {
 			Helper.New( ref s_content );
 			PopupWindow.Show( rect, s_content );
 		}
@@ -25,29 +27,25 @@ namespace HananokiEditor.CustomHierarchy {
 
 		/////////////////////////////////////////
 		public override Vector2 GetWindowSize() {
-			return new Vector2( 400, 300 );
+			return new Vector2( 400, EditorGUIUtility.singleLineHeight + ( EditorGUIUtility.singleLineHeight * Main.m_treeViewHideSelect.m_gameObjects.Length ) );
 		}
 
 
 		/////////////////////////////////////////
 		public override void OnGUI( Rect rect ) {
 			using( new GUILayout.AreaScope( rect ) )
-			using( var sc = new GUILayout.ScrollViewScope( m_scroll ) )
-			using( new GUILayoutScope( 16, 4 ) ) {
-				EditorGUIUtility.hierarchyMode = true;
-				EditorGUIUtility.wideMode = true;
-
-				m_scroll = sc.scrollPosition;
-				//m_currentEditor?.DrawDefaultInspector();
-				GUILayout.Space( 4 );
-				m_currentEditor?.OnInspectorGUI();
+			using( var sc = new GUILayout.ScrollViewScope( m_scroll ) ) {
+			//using( new GUILayoutScope( 4, 4 ) ) {
+				//GUILayout.Space(4);
+				Main.m_treeViewHideSelect.DrawLayoutGUI();
+				//GUILayout.Space( 4 );
 			}
 		}
 
 
 		/////////////////////////////////////////
 		public override void OnOpen() {
-			m_currentEditor = ComponentEditorCache.Get( s_component );
+			Main.m_treeViewHideSelect.RegisterFiles();
 		}