+++ Editor/Package.cs
@@ -7,6 +7,6 @@ namespace HananokiEditor.CustomHierarchy {
     public const string name = "CustomHierarchy";
     public const string nameNicify = "Custom Hierarchy";
     public const string editorPrefName = "Hananoki.CustomHierarchy";
-    public const string version = "0.8.0";
+    public const string version = "0.9.0";
   }
 }